Add - event participation limit todo : recettes

+++ app/Http/Controllers/EventParticipateController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Event;
 use App\Models\EventParticipates;
 use Illuminate\Http\Request;
 
@@ -15,11 +16,15 @@ class EventParticipateController extends Controller
         $existingParticipation = EventParticipates::where('events_id', $eventId)
             ->where('users_id', $userId)
             ->first();
-
-        if ($existingParticipation) {
+        $participantCount = EventParticipates::where('events_id', $eventId)->count();
+        if($participantCount >= Event::where('id', $eventId)->first()->max_participants){
+            return back()->with('error', 'The event is full.');
+        }
+        if ($existingParticipation ) {
             $existingParticipation->delete();
             return back();
         }
+
 
         // L'utilisateur ne participe pas encore à cet événement
         // Ajouter la participation à la table EventParticipates
